refactor(sponsor-reports): move report-errors helper to actions layer
